pathplannerlib-python/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robotpy-pathplannerlib"
version = "0.0.0"
description = "Python implementation of PathPlannerLib"
readme = "README.md"
authors = [{ name = "mjansen4857", email = "mjansen4857@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["pathplannerlib"]
dependencies = [
    "pyntcore >=2025.0.0b1",
    "robotpy-wpimath >=2025.0.0b1",
    "robotpy-commands-v2 >=2025.0.0b1",
    "robotpy-wpiutil >=2025.0.0b1",
    "robotpy-hal >=2025.0.0b1",
    "numpy >=2.1.1",
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/mjansen4857/pathplanner"